fix(main.js): extract MonacoEditor to a separate chunk [YTFRONT-3814]
