+++ packages/ui/src/ui/containers/AppNavigation/TopRowContent/TopRowContent.tsx
@@ -50,7 +50,7 @@ export default function TopRowContent() {
             {allowQT && (
                 <Route
                     path={`/:cluster/${Page.QUERIES}`}
-                    component={withLazyLoading(QueryTrackerTopRowLazy, null)}
+                    component={withLazyLoading(QueryTrackerTopRowLazy)}
                 />
             )}
             {hasOdinPage() && (
