+++ packages/ui/src/ui/hocs/withLazyLoading.tsx
@@ -1,14 +1,19 @@
 import React from 'react';
 
+import Loader from '../components/Loader/Loader';
+import ErrorBoundary from '../components/ErrorBoundary/ErrorBoundary';
+
 export default function withLazyLoading<P>(
     Component: React.ComponentType<P>,
     fallback?: React.ReactNode,
 ) {
     return function WithSuspense(props: P & React.JSX.IntrinsicAttributes) {
         return (
-            <React.Suspense fallback={fallback}>
-                <Component {...props} />
-            </React.Suspense>
+            <ErrorBoundary>
+                <React.Suspense fallback={fallback ?? <Loader visible centered size="l" />}>
+                    <Component {...props} />
+                </React.Suspense>
+            </ErrorBoundary>
         );
     };
 }
