--- packages/ui/src/ui/pages/query-tracker/QueryTracker/index.scss
@@ -1,6 +1,0 @@
-.query-tracker-page-lazy {
-    &__loader {
-        flex: 1 1 100%;
-        justify-content: center;
-    }
-}
